+++ algorithms/passive/ostia/src/main/java/de/learnlib/algorithms/ostia/Blue.java
@@ -1,21 +1,36 @@
+/* Copyright (C) 2013-2020 TU Dortmund
+ * This file is part of LearnLib, http://www.learnlib.de/.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package de.learnlib.algorithms.ostia;
 
 class Blue {
 
-    State parent;
-    int symbol;
+    final State parent;
+    final int symbol;
+
+    Blue(State parent, int symbol) {
+        this.symbol = symbol;
+        this.parent = parent;
+    }
 
     State state() {
         return parent.transitions[symbol].target;
     }
 
-    public Blue(State parent, int symbol) {
-        this.symbol = symbol;
-        this.parent = parent;
-    }
-
     @Override
     public String toString() {
-        return state().toString();
+        return String.valueOf(state());
     }
 }