+++ algorithms/passive/ostia/src/main/java/de/learnlib/algorithms/ostia/OSSTWrapper.java
@@ -1,15 +1,30 @@
+/* Copyright (C) 2013-2020 TU Dortmund
+ * This file is part of LearnLib, http://www.learnlib.de/.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package de.learnlib.algorithms.ostia;
 
 import java.util.ArrayDeque;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Queue;
 import java.util.Set;
 import java.util.stream.StreamSupport;
 
-import net.automatalib.automata.transducers.impl.compact.SubsequentialTransducer;
+import net.automatalib.automata.transducers.SubsequentialTransducer;
 import net.automatalib.words.Alphabet;
 import net.automatalib.words.Word;
 import net.automatalib.words.WordBuilder;
@@ -21,7 +36,7 @@ class OSSTWrapper<I, O> implements SubsequentialTransducer<State, I, Edge, O> {
     private final Alphabet<I> inputAlphabet;
     private final Alphabet<O> outputAlphabet;
 
-    public OSSTWrapper(State root, Alphabet<I> inputAlphabet, Alphabet<O> outputAlphabet) {
+    OSSTWrapper(State root, Alphabet<I> inputAlphabet, Alphabet<O> outputAlphabet) {
         this.root = root;
         this.inputAlphabet = inputAlphabet;
         this.outputAlphabet = outputAlphabet;
@@ -31,7 +46,7 @@ class OSSTWrapper<I, O> implements SubsequentialTransducer<State, I, Edge, O> {
     public Word<O> computeOutput(Iterable<? extends I> input) {
         final Iterator<Integer> transformedInput =
                 StreamSupport.stream(input.spliterator(), false).mapToInt(inputAlphabet).iterator();
-        final ArrayList<Integer> output = OSTIA.run(root, transformedInput);
+        final List<Integer> output = OSTIA.run(root, transformedInput);
 
         return output != null ?
                 output.stream().map(outputAlphabet::getSymbol).collect(Word.collector()) :
@@ -89,20 +104,21 @@ class OSSTWrapper<I, O> implements SubsequentialTransducer<State, I, Edge, O> {
         return transition.target;
     }
 
-    private Word<O> outToWord(Out out) {
+    private Word<O> outToWord(@Nullable Out out) {
         return outToWord(out == null ? null : out.str);
     }
 
-    private Word<O> outToWord(IntQueue out) {
+    private Word<O> outToWord(@Nullable IntQueue out) {
         if (out == null) {
             return Word.epsilon();
         }
 
         final WordBuilder<O> wb = new WordBuilder<>();
 
-        while (out != null) {
-            wb.add(outputAlphabet.getSymbol(out.value));
-            out = out.next;
+        IntQueue outIter = out;
+        while (outIter != null) {
+            wb.add(outputAlphabet.getSymbol(outIter.value));
+            outIter = outIter.next;
         }
 
         return wb.toWord();