+++ algorithms/passive/ostia/src/main/java/net/automatalib/automata/transducers/SSTVisualizationHelper.java
@@ -1,4 +1,19 @@
-package net.automatalib.automata.transducers.impl.compact;
+/* Copyright (C) 2013-2020 TU Dortmund
+ * This file is part of LearnLib, http://www.learnlib.de/.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package net.automatalib.automata.transducers;
 
 import java.util.Map;
 
@@ -35,7 +50,8 @@ public class SSTVisualizationHelper<S, I, T, O>
             return false;
         }
 
-        properties.put(NodeAttrs.LABEL, automaton.getStateProperty(node).toString());
+        final String oldLabel = properties.get(NodeAttrs.LABEL);
+        properties.put(NodeAttrs.LABEL, oldLabel + " / " + automaton.getStateProperty(node));
         return true;
     }
 }