+++ algorithms/passive/ostia/src/main/java/net/automatalib/automata/transducers/impl/compact/MutableOnwardSubsequentialTransducer.java
@@ -1,7 +1,0 @@
-package net.automatalib.automata.transducers.impl.compact;
-
-import net.automatalib.automata.MutableDeterministic;
-import net.automatalib.words.Word;
-
-public interface MutableOnwardSubsequentialTransducer<S, I, T, O> extends OnwardSubsequentialTransducer<S, I, T, O>,
-                                                                          MutableDeterministic<S, I, T, Word<O>, Word<O>> {}