+++ algorithms/passive/ostia/src/main/java/net/automatalib/automata/transducers/impl/compact/OnwardSubsequentialTransducer.java
@@ -1,3 +1,0 @@
-package net.automatalib.automata.transducers.impl.compact;
-
-public interface OnwardSubsequentialTransducer<S, I, T, O> extends SubsequentialTransducer<S, I, T, O> {}